requirements.txt
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
distlib==0.3.8
filelock==3.15.4
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prometheus_flask_exporter==0.23.1
psycopg2-binary==2.9.9
pytest==8.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
style==1.1.0
typing_extensions==4.12.2
update==0.0.1
virtualenv==20.26.3
Werkzeug==3.0.3
